fix movement and reset logic

+++ include/utility.h
@@ -3,6 +3,6 @@
 
 #include <string>
 #include <cstddef>
-int min(const int a, const size_t b){return (b<a) ? b : a;}
+int min(const int a, const size_t b){return (static_cast<int>(b)<a) ? b : a;}
 
 #endif
